move double diff
